drop dedicated files for the interfaces
can be moved into the file holding the default impl.
