+++ pkg/command/command.go
@@ -24,6 +24,26 @@ type Command struct {
 	exitCode             int
 }
 
+type runner interface {
+	SetDir(d string)
+	SetEnv(e []string)
+	Stdout(out io.Writer)
+	Stderr(err io.Writer)
+}
+
+// ExecRunner mock for intercepting calls to executables
+type ExecRunner interface {
+	runner
+	RunExecutable(e string, p ...string) error
+	RunExecutableInBackground(executable string, params ...string) (Execution, error)
+}
+
+// ShellRunner mock for intercepting shell calls
+type ShellRunner interface {
+	runner
+	RunShell(s string, c string) error
+}
+
 // SetDir sets the working directory for the execution
 func (c *Command) SetDir(d string) {
 	c.dir = d
